LPS-57210 move
