+++ modules/apps/mobile-device-rules/mobile-device-rules-service/src/com/liferay/mobile/device/rules/rule/group/action/BaseRedirectActionHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mobile.device.rules.rulegroup.action;
+package com.liferay.mobile.device.rules.rule.group.action;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
